Update to Burst animation Trigger
Created a starburst animation next to Memory Collector in that object in Unity. Then, I created a state called empty and connected entry to empty then to Burst then to empty again. I changed the alpha on the sprite to 0 so when it comes in it will fade in to 1 then fade out. I created a trigger named Burst. I went to the Inventory script under the PuzzleInventoryCollector object? and I added Animator to the script and added my code to set the trigger up. Due to compiling package errors, I am not able to add my object to the inventory in the inspector which means it won't be seen. My mentor said this might be because I do not have a package downloaded, or something was not pulled.